refactor: move 'combiner' tests to test_paradigm_filler.py
refactor: move 'combiner' tests to test_paradigm_filler.py

remove: test_layout_combiner.py
